refactor cusomui development environment
